rename corgi handler module
