+++ openlcsd/flow/tasks.py
@@ -15,7 +15,7 @@ from packagedcode.maven import parse as maven_parse
 from openlcsd.celery import app
 from openlcsd.flow.task_wrapper import WorkflowWrapperTask
 from openlcs.libs.common import get_nvr_list_from_components
-from openlcs.libs.corgi_handler import ParentComponentsAsync
+from openlcs.libs.corgi import ParentComponentsAsync
 from openlcs.libs.driver import OpenlcsClient
 from openlcs.libs.kojiconnector import KojiConnector
 from openlcs.libs.logger import get_task_logger
